+++ index.js
@@ -13,6 +13,7 @@ import inatjs from "inaturalistjs";
 import INatPaperProvider from "providers/INatPaperProvider";
 import ObsEditProvider from "providers/ObsEditProvider";
 import RealmProvider from "providers/RealmProvider";
+import UploadProgressProvider from "providers/UploadProgressProvider";
 import React from "react";
 import { AppRegistry } from "react-native";
 import Config from "react-native-config";
@@ -20,7 +21,6 @@ import { setJSExceptionHandler, setNativeExceptionHandler } from "react-native-e
 import { GestureHandlerRootView } from "react-native-gesture-handler";
 import { startNetworkLogging } from "react-native-network-logger";
 import { SafeAreaProvider } from "react-native-safe-area-context";
-import { UploadProgressProvider } from "sharedHooks/useUploadProgress";
 
 import { name as appName } from "./app.json";
 import { log } from "./react-native-logs.config";
